feat: refactor hover icons and auth error messages
