+++ frontend/errors/auth-messages.ts
@@ -1,0 +1,8 @@
+enum AuthErrorEnumMessages {
+  DEFAULT = "An error occurred. Please check your internet connection.",
+  EMAIL_EXISTS = "Registration failed. This email is already registered. Please try to login",
+  INVALID_CREDENTIALS = "Login failed. Please, make sure you entered the right credentials",
+  USER_NOT_FOUND = "Login failed. Please, make sure you are already registered",
+}
+
+export default AuthErrorEnumMessages;
