+++ frontend/utils/errorCodes.ts
@@ -1,0 +1,2 @@
+export const errorCodes = (message: string): number =>
+  Number(message.slice(message.length - 3, message.length));
